Move `forPublishingDocumentation` factory method to its context
